Add chakra.pxd file.     allows to use chakra module from others cython module. pyx and pyx.in files adapted for the use of the pxd file

+++ chakra/clean.py
@@ -23,7 +23,7 @@ FILE=['wrapper_chakra_obj', 'wrapper_chakra_host_obj','wrapper_magma']
 if os.path.exists('wrapper_chakra.cpp'):
     os.remove('wrapper_chakra.cpp')
 for f in FILE:
-    if os.path.exists(f+'.pyx'):
+    if (os.path.exists(f+'.pyx') and os.path.exists(f+'.pyx.in') ) :
         os.remove(f+'.pyx')
     if os.path.exists(f+'.cpp'):
         os.remove(f+'.cpp')
